ProductSeeder (se lleno la tabla con los productos disponibles)

+++ database/migrations/2018_01_26_020257_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('NombreProducto');
             $table->integer('category_id')->unsigned();
             $table->foreign('category_id')->references('id')->on('category_products')->onupdate('cascade');
-            $table->integer('Cantidad');
+            $table->float('Cantidad', 8, 2);
             $table->integer('merma');
             $table->timestamps();
         });
